Combine tool and registry files into tool_registry.rs
